Add WinGet support for listing installed packages
